Revert "Revert "refactor: move report`s modules into report folder""
This reverts commit 456173b856232ced5f47ccf1bf53ebfbac83247d.
