+++ src/auth/auth.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { LocalStrategy } from './strategies/local.strategy';
 import { JwtStrategy, RefreshTokenStrategy } from './strategies/jwt.strategy';
@@ -20,7 +19,6 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 @Module({
   imports: [
     TypeOrmModule.forFeature([User]),
-    PassportModule,
     UserModule,
     MailModule,
     JwtModule.registerAsync({
