+++ src/report/normalizeReports/types/types/freedom-finance.type.ts
