+++ src/report/normalizeReports/types/types/freedomFinance.type.ts
