+++ src/report/normalizeReports/types/types/stock-exchange.type.ts
@@ -1,4 +1,4 @@
-import { StockExchangeEnum } from 'src/normalizeTrades/constants/enums';
+import { StockExchangeEnum } from 'src/report/normalizeTrades/constants/enums';
 
 export type StockExchangeType =
   | StockExchangeEnum.FREEDOM_FINANCE
