+++ src/report/normalizeReports/types/types/trades-by-stock-exchange.type.ts
