+++ src/report/normalizeTrades/constants/enums.ts
