+++ src/report/normalizeTrades/types/types/trades-by-stock-exchange.type.ts
