+++ src/report/trade/trade.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { TradeService } from './trade.service';
 
 @Module({
-  imports: [],
   exports: [TradeService],
   providers: [TradeService],
 })
