+++ src/user/user.module.ts
@@ -6,7 +6,7 @@ import { User } from './entities/user.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
-  exports: [TypeOrmModule, UserService],
+  exports: [UserService],
   providers: [UserService],
   controllers: [UserController],
 })
